requirements.txt

python-dotenv==1.0.0
openai
streamlit
PyPDF2
fpdf